+++ R/zzz.R
@@ -1,0 +1,4 @@
+.onLoad <- function(libname, pkgname) {
+  # TODO after https://github.com/extendr/extendr/issues/278 resolved, we should remove this
+  set_panic_hook()
+}