+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `ymd()` now supports `...` arguments, which is convinient for interactive use, e.g., `ymd(210101, 220201)`.
 * Implement date part extracting functions, including `year()`, `month()`, `quarter()`, `iso_week()`, `iso_wday()`, `wday()`, `mday()`, `yday()`,
+* Redirect the Rust message to R's stderr stream.
 
 # ymd 0.0.1